# config/zyrquen-openapi-spec.yaml
openapi: 3.0.3
info:
  title: ZYRQUEN Ω∞ Sovereign Gateway & Chamber 17 Audit Services API
  description: 'OpenAPI 3.0.3 Specification for ZYRQUEN Ω∞ Sovereign Control Plane,
    Chamber 17 Audit Services, and Security Interceptors.


    ### Legal & Technical Compliance Framework:

    - **Thai Electronic Transactions Act B.E. 2544 (2001)**: Sections 9 (e-Signatures),
    26 (Advanced Secure Signatures & Non-repudiation), 28 (CA Certificates & Public
    Verification)

    - **Thai Personal Data Protection Act B.E. 2562 (2019)**: Section 37 (Zero-Knowledge
    Privacy Isolation Vault & PII Scrubbing)

    - **Hardware Security**: 10/10 REAL_HSM Quorum (FIPS 140-3 Level 4 / CC EAL6+)
    with Active Zeroization (< 1.2ms)

    - **Post-Quantum Cryptography**: CRYSTALS-Dilithium-5 (ML-DSA-87 / FIPS 204),
    SPHINCS+ (SLH-DSA-192 / FIPS 205)

    - **SSoT Invariant**: Genesis Block #849202, Merkle Root `0x909ab814479844d8a14816bed34cdbb07528e18501da86fc4691763a43fa4c68`,
    Zero Drift (0.00%)'
  version: 1.2.0-LTS
  contact:
    name: Yuttaphum Phakphian (Sovereign Principal Architect)
    url: https://github.com/yuththaphum-phakphian/ZYRQUEN-1.2-LTS
    email: sovereign@zyrquen.internal
  license:
    name: LOCKED_FROZEN_v1.2_LTS Sovereign License
servers:
- url: https://nexus.zyrquen.internal/v1
  description: Bangkok Root Citadel Server (BK01)
- url: https://sg02.zyrquen.internal/v1
  description: Singapore Relay Node (SG02)
- url: http://localhost:3000
  description: Local Sovereign Gateway Simulation Environment
security:
- PqcSignatureAuth: []
paths:
  /api/v1/telemetry:
    get:
      summary: Get Real-time Sub-Kelvin Telemetry & SSoT Baseline
      description: Returns live system telemetry including Cryo temperature (14.98
        mK), QOps throughput (851.9 QOps/s), quantum coherence, and verifies the Genesis
        Merkle Root invariant.
      operationId: getTelemetry
      tags:
      - Telemetry & Homeostasis
      security: []
      responses:
        '200':
          description: Telemetry stream successfully retrieved with SSoT Δ0 invariant
            verified.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TelemetryResponse'
              example:
                status: SUCCESS
                systemStatus: LOCKED_FROZEN_v1.2_LTS
                blockHeight: 849202
                merkleGenesis: '0x909ab814479844d8a14816bed34cdbb07528e18501da86fc4691763a43fa4c68'
                cryoTempMK: 14.98
                qopsThroughput: 851.9
                coherencePct: 99.992
                zeroDrift: 0.00%
  /api/v1/audit/records:
    get:
      summary: Query Immutable Audit Evidence Records
      description: Retrieves unalterable digital evidence records from Chamber 17
        and Module 17 V24 WORM Storage (Zero-Deletion Guarantee).
      operationId: getAuditRecords
      tags:
      - Audit & Evidence
      parameters:
      - name: sealId
        in: query
        required: true
        description: Canonical Seal ID (1 to 14,902)
        schema:
          type: integer
          example: 14902
      responses:
        '200':
          description: Canonical audit record found and verified intact.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuditRecordResponse'
              example:
                sealId: 14902
                blockHeight: 849202
                merkleLeafHash: '0x5a13396c129c611f15232fdaf54bfad00c4147abdbc3424c71e4ec103dcc8cc3'
                status: VERIFIED_INTACT
                wormStorage: Module 17 V24 WORM
                legalTag: ETA B.E. 2544 Sec 28 / Delete-Nothing Enforced
        '401':
          description: Unauthorized. Requires IAL2+/AAL2+ e-signature headers.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /api/v1/audit/replay:
    post:
      summary: Execute 12-Stage Forensic Trace Replay
      description: Triggers sequential 12-stage forensic trace simulation for evidence
        verification within SLA < 142.00ms.
      operationId: executeTraceReplay
      tags:
      - Forensics & Replay
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/TraceReplayRequest'
            example:
              sealId: 14903
              auditMode: FULL_TRACE_REPLAY
      responses:
        '200':
          description: 12-Stage trace replay executed successfully within SLA bounds.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TraceReplayResponse'
              example:
                sealId: 14903
                status: COMPLETED
                executionTimeMs: 35.8
                slaLimitMs: 142.0
                verdict: 100% COURT-ADMISSIBLE READY
                stagesPassed: 12
                finalStage: 'STAGE-12: CLOSURE (Immutable WORM Finalized)'
  /api/v2/auth/register:
    post:
      summary: User Identity & Compliance Registration (Section 26 Gate)
      description: Evaluates user IAL/AAL compliance tier, runs Sentinel AI Risk Interceptor,
        and registers identity with Dilithium-5 signature headers.
      operationId: registerUserAuth
      tags:
      - Identity & Governance
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RegisterAuthRequest'
            example:
              user:
                id: USR-001
                name: Yuttaphum Phakphian
                role: Sovereign Principal Architect
              auth:
                ial: 3
                aal: 3
                cryptoScheme: Dilithium-5
                hsmSigned: true
      responses:
        '200':
          description: User identity successfully verified under ETA B.E. 2544 Section
            26.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RegisterAuthResponse'
              example:
                status: SUCCESS
                system_status: LOCKED_FROZEN_v1.2_LTS
                verdict: APPROVED_SECTION_26
                reason: Passed Section 26 compliance. Advanced digital signature ensures
                  integrity & non-repudiation.
                sentinel_risk_score: 0.02
                user_profile:
                  id: USR-001
                  name: Yuttaphum Phakphian
                  role: Sovereign Principal Architect
                  registered_at: '2026-09-21T07:18:13Z'
        '403':
          description: Quarantined! Sentinel AI Interceptor detected risk score >=
            0.85.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/QuarantineErrorResponse'
              example:
                error: ZYRQUEN_QUARANTINE_TRIGGERED
                verdict: QUARANTINED
                chamber: Chamber 02 (FORENSICS & QUARANTINE)
                riskScore: 0.96
                reason: Risk score (0.96) exceeds threshold (0.85). Isolated to Chamber
                  02.
                timestamp: '2026-09-21T07:18:13Z'
  /api/v2/treasury/refund:
    post:
      summary: FIOS Treasury Gas Refund Allocation (Section 28 Sovereign Gate)
      description: Allocates gas refund from the ฿12.5M pool using the Nc x Vc Chain
        Model of Segment Value. Requires 10/10 REAL_HSM Quorum ratification and Dilithium-5
        signatures.
      operationId: allocateTreasuryRefund
      tags:
      - Sovereign Treasury
      security:
      - HsmQuorumAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/TreasuryRefundRequest'
            example:
              allocationSegment: Gen_Z_Core
              totalGasRefundPoolThb: 12500000.0
      responses:
        '200':
          description: Treasury gas refund allocated with Zero Drift 0.00% verified.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TreasuryRefundResponse'
              example:
                status: COMPLETED
                verdict: APPROVED_SECTION_28
                reason: CA-Certified secure signature bound to 10/10 REAL_HSM Quorum
                  (FIPS 140-3 Level 4).
                genesis_block: 849202
                merkle_root: '0x909ab814479844d8a14816bed34cdbb07528e18501da86fc4691763a43fa4c68'
                audit_trail:
                  zero_drift: 0.00%
                  integrity: VERIFIED_MODULE_17
                  thai_law_compliance: พ.ร.บ. ว่าด้วยธุรกรรมทางอิเล็กทรอนิกส์ พ.ศ.
                    2544 มาตรา 28
                distribution:
                  segment: Gen_Z_Core
                  segment_market_value_thb: 134400000.0
                  allocated_gas_refund_thb: 1179709.01
                  per_capita_refund_thb: 0.08778
                  pqc_signature: SIG_PQC_DILITHIUM-5_BC2B1C7991D05470_10/10_REAL_HSM_RATIFIED
        '401':
          description: Unauthorized. 10/10 REAL_HSM Quorum or Dilithium-5 signature
            missing.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              example:
                error: UNAUTHORIZED_SECTION_28_REJECTED
                reason: Sovereign treasury functions require IAL2+, AAL2+, Dilithium-5
                  and 10/10 REAL_HSM signature verification.
  /api/v1/gold-seal/verify:
    post:
      summary: Public Merkle Proof & Gold Seal Judicial Audit
      description: Public verification endpoint for courts, law enforcement, and ETDA
        auditors to verify Merkle Leaf inclusion against the Genesis Merkle Root.
      operationId: verifyGoldSeal
      tags:
      - Public Verification & Judicial Audit
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/GoldSealVerifyRequest'
            example:
              sealId: 14902
              leafHash: '0x5a13396c129c611f15232fdaf54bfad00c4147abdbc3424c71e4ec103dcc8cc3'
      responses:
        '200':
          description: 'Merkle proof verified intact against Block #849202 Genesis
            Root.'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/GoldSealVerifyResponse'
              example:
                verified: true
                blockHeight: 849202
                merkleRoot: '0x909ab814479844d8a14816bed34cdbb07528e18501da86fc4691763a43fa4c68'
                zeroDrift: 0.00%
                courtAdmissibility: 100% COURT-ADMISSIBLE READY
components:
  securitySchemes:
    PqcSignatureAuth:
      type: apiKey
      in: header
      name: X-Zyrquen-Sovereign-Sig
      description: Post-Quantum Dilithium-5 (ML-DSA-87) signature header bound to
        user identity and payload hash.
    HsmQuorumAuth:
      type: apiKey
      in: header
      name: X-Zyrquen-HSM-Quorum
      description: Unanimous 10/10 REAL_HSM Quorum attestation token (FIPS 140-3 Level
        4).
  schemas:
    TelemetryResponse:
      type: object
      properties:
        status:
          type: string
          example: SUCCESS
        systemStatus:
          type: string
          example: LOCKED_FROZEN_v1.2_LTS
        blockHeight:
          type: integer
          example: 849202
        merkleGenesis:
          type: string
          example: '0x909ab814479844d8a14816bed34cdbb07528e18501da86fc4691763a43fa4c68'
        cryoTempMK:
          type: number
          format: float
          example: 14.98
        qopsThroughput:
          type: number
          format: float
          example: 851.9
        coherencePct:
          type: number
          format: float
          example: 99.992
        zeroDrift:
          type: string
          example: 0.00%
    AuditRecordResponse:
      type: object
      properties:
        sealId:
          type: integer
          example: 14902
        blockHeight:
          type: integer
          example: 849202
        merkleLeafHash:
          type: string
          example: '0x5a13396c129c611f15232fdaf54bfad00c4147abdbc3424c71e4ec103dcc8cc3'
        status:
          type: string
          example: VERIFIED_INTACT
        wormStorage:
          type: string
          example: Module 17 V24 WORM
        legalTag:
          type: string
          example: ETA B.E. 2544 Sec 28 / Delete-Nothing Enforced
    TraceReplayRequest:
      type: object
      required:
      - sealId
      properties:
        sealId:
          type: integer
          example: 14903
        auditMode:
          type: string
          example: FULL_TRACE_REPLAY
    TraceReplayResponse:
      type: object
      properties:
        sealId:
          type: integer
          example: 14903
        status:
          type: string
          example: COMPLETED
        executionTimeMs:
          type: number
          format: float
          example: 35.8
        slaLimitMs:
          type: number
          format: float
          example: 142.0
        verdict:
          type: string
          example: 100% COURT-ADMISSIBLE READY
        stagesPassed:
          type: integer
          example: 12
        finalStage:
          type: string
          example: 'STAGE-12: CLOSURE (Immutable WORM Finalized)'
    RegisterAuthRequest:
      type: object
      required:
      - user
      - auth
      properties:
        user:
          type: object
          properties:
            id:
              type: string
              example: USR-001
            name:
              type: string
              example: Yuttaphum Phakphian
            role:
              type: string
              example: Sovereign Principal Architect
        auth:
          type: object
          properties:
            ial:
              type: integer
              example: 3
            aal:
              type: integer
              example: 3
            cryptoScheme:
              type: string
              example: Dilithium-5
            hsmSigned:
              type: boolean
              example: true
    RegisterAuthResponse:
      type: object
      properties:
        status:
          type: string
          example: SUCCESS
        system_status:
          type: string
          example: LOCKED_FROZEN_v1.2_LTS
        verdict:
          type: string
          example: APPROVED_SECTION_26
        reason:
          type: string
        sentinel_risk_score:
          type: number
          format: float
          example: 0.02
        user_profile:
          type: object
    TreasuryRefundRequest:
      type: object
      required:
      - allocationSegment
      properties:
        allocationSegment:
          type: string
          enum:
          - Gen_Z_Core
          - Gen_Y_Pro
          - Gen_X_Enterprise
          - SMB_Retail
          example: Gen_Z_Core
        totalGasRefundPoolThb:
          type: number
          format: float
          example: 12500000.0
    TreasuryRefundResponse:
      type: object
      properties:
        status:
          type: string
          example: COMPLETED
        verdict:
          type: string
          example: APPROVED_SECTION_28
        reason:
          type: string
        genesis_block:
          type: integer
          example: 849202
        merkle_root:
          type: string
        audit_trail:
          type: object
        distribution:
          type: object
    GoldSealVerifyRequest:
      type: object
      required:
      - sealId
      - leafHash
      properties:
        sealId:
          type: integer
          example: 14902
        leafHash:
          type: string
          example: '0x5a13396c129c611f15232fdaf54bfad00c4147abdbc3424c71e4ec103dcc8cc3'
    GoldSealVerifyResponse:
      type: object
      properties:
        verified:
          type: boolean
          example: true
        blockHeight:
          type: integer
          example: 849202
        merkleRoot:
          type: string
        zeroDrift:
          type: string
          example: 0.00%
        courtAdmissibility:
          type: string
          example: 100% COURT-ADMISSIBLE READY
    QuarantineErrorResponse:
      type: object
      properties:
        error:
          type: string
          example: ZYRQUEN_QUARANTINE_TRIGGERED
        verdict:
          type: string
          example: QUARANTINED
        chamber:
          type: string
          example: Chamber 02 (FORENSICS & QUARANTINE)
        riskScore:
          type: number
          format: float
          example: 0.96
        reason:
          type: string
        timestamp:
          type: string
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
        reason:
          type: string